refactor planner hooks
