+++ src/components/planner/DayCard.tsx
@@ -10,11 +10,11 @@
 import "./style.css";
 import * as React from "react";
 import {
-  useDay,
   useSelectedProject,
   useSelectedTask,
   type ISODate,
 } from "./usePlanner";
+import { useDay } from "./useDay";
 import Input from "@/components/ui/primitives/Input";
 import { cn } from "@/lib/utils";
 import DayCardHeader from "./DayCardHeader";
