+++ src/components/planner/WeekPicker.tsx
@@ -11,7 +11,8 @@
 import * as React from "react";
 import Hero2 from "@/components/ui/layout/Hero2";
 import Button from "@/components/ui/primitives/Button";
-import { useFocusDate, useDay, type ISODate } from "./usePlanner";
+import { useFocusDate, type ISODate } from "./usePlanner";
+import { useDay } from "./useDay";
 import { cn } from "@/lib/utils";
 import { CalendarDays, ChevronLeft, ChevronRight, ArrowUpToLine } from "lucide-react";
 import { isoToDate, toISO, addDays, mondayStartOfWeek } from "@/lib/date";
