+++ src/components/planner/index.ts
@@ -15,4 +15,7 @@ export { default as WeekNotes } from "./WeekNotes";
 export { default as WeekPicker } from "./WeekPicker";
 export { default as WeekSummary } from "./WeekSummary";
 export * from "./dayCrud";
+export * from "./plannerCrud";
+export * from "./plannerStore";
+export * from "./useDay";
 export * from "./usePlanner";
